@@ -4,7 +4,7 @@
   "description": "A blank schematics",
   "scripts": {
     "build": "tsc -p tsconfig.json",
-    "test": "jest --coverage",
+    "test": "jest --coverage --env=node",
     "lint": "eslint ./src --ext .ts",
     "preinstall": "npx npm-force-resolutions",
     "postbuild": "bash package.sh",
@@ -23,36 +23,34 @@
   "license": "MIT",
   "schematics": "./src/collection.json",
   "dependencies": {
-    "@angular-devkit/core": "^12.0.0",
-    "@angular-devkit/schematics": "^12.0.0",
+    "@angular-devkit/core": "^13.3.2",
+    "@angular-devkit/schematics": "^13.3.2",
+    "axios": "^0.26.1",
     "find-up": "^5.0.0",
-    "imng-ngrx-utils": "^3.291.2",
     "lodash": "^4.17.21",
-    "node-fetch": "^2.6.5",
     "pluralize": "^8.0.0",
     "rxjs": "^6.6.7"
   },
   "devDependencies": {
-    "@angular-devkit/schematics-cli": "^12.0.0",
-    "@angular/compiler": "^12.2.7",
-    "@angular/core": "^12.2.7",
-    "@nrwl/angular": "^12.9.0",
-    "@types/jest": "^26.0.24",
-    "@types/lodash": "^4.14.175",
-    "@types/node": "^16.10.2",
-    "@types/node-fetch": "^2.5.12",
+    "@angular-devkit/schematics-cli": "^13.3.2",
+    "@angular/compiler": "^13.3.2",
+    "@angular/core": "^13.3.2",
+    "@types/jest": "^27.4.1",
+    "@types/lodash": "^4.14.181",
+    "@types/node": "^16.11.26",
     "@types/pluralize": "0.0.29",
-    "@typescript-eslint/eslint-plugin": "^4.13.0",
-    "@typescript-eslint/parser": "^4.13.0",
+    "@typescript-eslint/eslint-plugin": "^5.19.0",
+    "@typescript-eslint/parser": "^5.19.0",
     "codelyzer": "^6.0.1",
-    "eslint": "^7.17.0",
-    "eslint-config-prettier": "8.3",
-    "jest": "^26.6.3",
-    "jest-junit": "^12.0.0",
+    "eslint": "^8.13.0",
+    "eslint-config-prettier": "8.5",
+    "imng-ngrx-utils": "4.89",
+    "jest": "^27.5.1",
+    "jest-junit": "^13.1.0",
     "npm-force-resolutions": "^0.0.10",
     "schematics": "^0.2.0",
-    "snyk": "^1.742.0",
-    "ts-jest": "^26.4.4",
+    "snyk": "^1.904.0",
+    "ts-jest": "^27.1.4",
     "typescript": "^4.4.4"
   }
 }
